Make Rust the policy and test safety net for operator config and CLIs.
Port deleted pytest coverage into Rust integration tests, add rigorous
markets/program validation, route scripts through config-validate, and
extend CI/pre-commit with cargo fmt/clippy/test on arm and macOS.

+++ greenfloor-engine/src/daemon/inventory_phase.rs
@@ -70,7 +70,7 @@ pub async fn run_inventory_phase(
         return Ok(BTreeMap::new());
     }
 
-    let base_unit_multiplier = default_mojo_multiplier_for_asset(market.base_asset.trim()) as i64;
+    let base_unit_multiplier = default_mojo_multiplier_for_asset(market.base_asset.trim());
     let scan_result: SignerResult<(String, usize, BTreeMap<i64, i64>)> = async {
         let signer_config = load_signer_config(resources.program_path())?;
         let (resolved_base_asset_id, _) =
